requirements.txt
Telethon~=1.38.1
python-dotenv